src/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.38.0
colorama==0.4.6
distro==1.9.0
dnspython==2.6.1
ecdsa==0.18.0
fastapi==0.109.2
h11==0.14.0
httpcore==1.0.3
httptools==0.6.1
httpx==0.26.0
idna==3.6
mangum==0.17.0
motor==3.3.2
openai==1.12.0
passlib==1.7.4
pyasn1==0.5.1
pydantic==2.6.1
pydantic_core==2.16.2
pymongo==4.6.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytube==15.0.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
whisper==1.1.10
youtube-transcript-api==0.6.2
pypdf
langchain_community
langchain-text-splitters
langchain-openai
langchain
boto3
yt_dlp
langchain-chroma
